feat(server): add getServerSession() & getServerProviders()
